Made improvements to settings page components.  Refactored and made code improvements.
