+++ fervent-web/src/components/buttons/Button.jsx
@@ -6,24 +6,22 @@ import Typography from "@mui/material/Typography";
  * Button.
  * @description A generic all-purpose button component.
  * @param {object} props
- * @param {string} props.title Button title.
- * @param {React.JSX.Element} props.icon Button icon.
- * @param {"contained" | "text" | "outlined"} props.variant Button variant.
- * @param {boolean} props.autoFocus AutoFocus.
- * @param {Function} props.handleClick OnClick event handler.
  * @returns JSX Component.
  */
-function Button({ title, icon, variant="contained", autoFocus = false, handleClick }) {
+function Button({ title, handleClick, icon, variant = "contained", color = "primary", autoFocus = false, loading = false,
+    loadingPosition = "center", sx = {} }) {
     return (
         <MuiButton
-            color="primary"
-            variant={variant}
-            size="small"
             type="button"
-            component="label"
+            size="small"
+            color={color}
+            variant={variant}
             onClick={handleClick}
             startIcon={icon}
-            autoFocus
+            autoFocus={autoFocus}
+            loading={loading}
+            loadingPosition={loadingPosition}
+            sx={sx}
         >
             <Typography textTransform={"none"}>{title}</Typography>
         </MuiButton>
