+++ fervent-web/src/components/settings/components/TabList.jsx
@@ -1,0 +1,14 @@
+import GeneralSettings from "./general/GeneralSettings";
+
+/**
+ * Settings tabs.
+ */
+const Tabs = [
+    {
+        id: 0,
+        title: "General",
+        element: <GeneralSettings />,
+    }
+];
+
+export default Tabs;
